Rename project
Renames the project from `labeler` to `labelr`.

The references in the package and the documentation were changed as
well.
