+++ cmd/apply.go
@@ -1,31 +1,10 @@
-/*
-Copyright © 2019 Rémy Greinhofer <remy.greinhofer@gmail.com>
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
-*/
 package cmd
 
 import (
 	"log"
 	"os"
 
-	"github.com/rgreinho/labeler/labeler"
+	"github.com/rgreinho/labelr/labelr"
 	"github.com/spf13/cobra"
 )
 
@@ -40,7 +19,7 @@ var applyCmd = &cobra.Command{
 		// 1. from repo
 		// 2. from env var
 		// 3. from CLI
-		owner, repository := labeler.GetInfo()
+		owner, repository := labelr.GetInfo()
 		if owner == "" {
 			o, err := cmd.Flags().GetString("owner")
 			if err != nil {
@@ -77,7 +56,7 @@ var applyCmd = &cobra.Command{
 		}
 
 		// Prepare client.
-		g := labeler.NewLabeler(owner, repository, token)
+		g := labelr.NewLabelr(owner, repository, token)
 
 		// Get sync flag (delete labels if need be).
 		sync, _ := cmd.Flags().GetBool("sync")
