+++ go.mod
@@ -1,15 +1,11 @@
-module github.com/rgreinho/labeler
+module github.com/rgreinho/labelr
 
 go 1.12
 
 require (
-	github.com/golang/protobuf v1.3.1 // indirect
 	github.com/google/go-github/v28 v28.0.1
-	github.com/kr/pretty v0.1.0 // indirect
 	github.com/spf13/cobra v0.0.5
-	golang.org/x/net v0.0.0-20190522155817-f3200d17e092 // indirect
 	golang.org/x/oauth2 v0.0.0-20180821212333-d2e6202438be
-	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 	gopkg.in/yaml.v1 v1.0.0-20140924161607-9f9df34309c0
 	gopkg.in/yaml.v2 v2.2.2
 )
